Makes reset button recreate the baord
